face-emotion-detector/requirements.txt
# Face Emotion and Gender Detector Requirements
opencv-python>=4.5.0
mediapipe>=0.8.9
numpy>=1.19.0
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.0.0